requirements.txt
openai-whisper
sounddevice
scipy
scikit-learn
datasets
joblib
spotipy
streamlit
